requirements.txt
# 核心依赖
ultralytics>=8.0.0        # YOLOv8/v9/v10
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0      # HuggingFace Transformers

# 数据处理
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0     # 数据增强

# 数据集工具
roboflow>=1.0.0           # 数据集管理
label-studio>=1.0.0       # 标注工具 (可选)

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 工具库
tqdm>=4.65.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# 部署相关 (可选)
onnx>=1.14.0              # 模型导出
onnxruntime>=1.15.0       # ONNX 推理
tensorrt>=8.6.0           # TensorRT 加速 (NVIDIA GPU)

# Web 演示 (可选)
gradio>=4.0.0             # 快速 Demo
streamlit>=1.28.0         # Web 应用
